Implemented cache-efficient split dentries.
To conserve CPU cache on lookups, the directory entry structure is
split into two parts. The hot section (struct dentry) is the inum and
a hash for the string name; the cold section (struct dentry_ext) has
the full name. Since both sections are fixed-sized, the cold section's
location can be obtained by calculating the offset.

Additionally, a new structure (struct dentry_block) is added to reflect
the on-disk layout of a single 4KB directory block. (fsemu/include/fs.h)

When performing a lookup, multiple dentries (hot section) can fit into
one cache line, only when we have a potential candidate (hash value
checks out) do we bring in the cold section to check the full name.

A shorter stride also allows CPU prefetching to perform better, although
this may not be evident when benchmarking using cachegrind, since it is
not clear how prefetching is simulated in cachegrind.

Because the structure is not split into two parts, we can no longer rely
on the simple C sytnax to access members (e.g. dent->name), so a series
of inlined getter functions are provided.

As a result of this change, there is no longer a dentry entity inside
the superblock. Instead, the file system now relies on the fact that the
root inode will always be present at index 1.