Adjusted visibility of kernel data structures
Moved the openfiles[] array (file descriptors) to the kernel side --
they are no longer exposed to the "user level" by being in fsemu.h.

This means that fsemu.h no longer has a need to include any of our
custom header files. Ideally, the only header file ANY "user-level" file
should include is "fs_syscall.h".

As a result of this readjustment of dependencies, changes had to be made
to several other files.