. and .. entries, more special cases.
1. Inline directory lookups
Before this update, lookup paths containing . and .. entries of inlined
directories could not be resolved since the lookup function could not
actually locate these entries in inlined directories. Now special
handling is added for this special scenario.

2. Renaming/deleting . and .. entries
Renaming . and .. is now prohibitted in the system call function, as
well as any attempt to unlink these dentries.